test: testing put only limited s3 policy
