+++ cpp/src/arrow/dataset/dataset_writer.cc
@@ -459,9 +459,11 @@ class DatasetWriter::DatasetWriterImpl : public util::AsyncDestroyable {
     if (!directory.empty()) {
       auto full_path =
           fs::internal::ConcatAbstractPath(write_options_.base_dir, directory);
-      return DoWriteRecordBatch(std::move(batch), full_path, prefix, write_options_.create_dir);
+      return DoWriteRecordBatch(std::move(batch), full_path, prefix,
+                                write_options_.create_dir);
     } else {
-      return DoWriteRecordBatch(std::move(batch), write_options_.base_dir, prefix,  write_options_.create_dir);
+      return DoWriteRecordBatch(std::move(batch), write_options_.base_dir, prefix,
+                                write_options_.create_dir);
     }
   }
 
@@ -480,7 +482,8 @@ class DatasetWriter::DatasetWriterImpl : public util::AsyncDestroyable {
   }
 
   Future<> DoWriteRecordBatch(std::shared_ptr<RecordBatch> batch,
-                              const std::string& directory, const std::string& prefix, const bool& create_dir) {
+                              const std::string& directory, const std::string& prefix,
+                              const bool& create_dir) {
     ARROW_ASSIGN_OR_RAISE(
         auto dir_queue_itr,
         ::arrow::internal::GetOrInsertGenerated(
@@ -519,7 +522,7 @@ class DatasetWriter::DatasetWriterImpl : public util::AsyncDestroyable {
     }
 
     if (batch) {
-      return backpressure.Then([this, batch, directory, prefix] {
+      return backpressure.Then([this, batch, directory, prefix, create_dir] {
         return DoWriteRecordBatch(batch, directory, prefix, create_dir);
       });
     }
