+++ python/pyarrow/tests/conftest.py
@@ -16,6 +16,7 @@
 # under the License.
 
 import os
+import sys
 import pathlib
 import subprocess
 from tempfile import TemporaryDirectory
@@ -25,6 +26,7 @@ import hypothesis as h
 
 from pyarrow.util import find_free_port
 from pyarrow import Codec
+from pyarrow.tests.util import _configure_limited_user
 
 
 # setup hypothesis profiles
@@ -311,3 +313,23 @@ def s3_server(s3_connection):
         finally:
             if proc is not None:
                 proc.kill()
+
+
+@pytest.fixture(scope='session')
+def limited_s3_user(request, s3_server):
+    def _limited_s3_user(policy):
+        if sys.platform == 'win32':
+            # Can't rely on FileNotFound check because
+            # there is sometimes an mc command on Windows
+            # which is unrelated to the minio mc
+            pytest.skip('The mc command is not installed on Windows')
+        request.config.pyarrow.requires('s3')
+        tempdir = s3_server['tempdir']
+        host, port, access_key, secret_key = s3_server['connection']
+        address = '{}:{}'.format(host, port)
+        if not _configure_limited_user(tempdir, address,
+                                       access_key, secret_key, policy):
+            pytest.skip(
+                'Could not locate mc command to configure limited user')
+
+    return _limited_s3_user
